+++ src/main/java/me/alpha432/oyvey/features/commands/Command.java
@@ -3,10 +3,7 @@ package me.alpha432.oyvey.features.commands;
 import me.alpha432.oyvey.OyVey;
 import me.alpha432.oyvey.features.Feature;
 import me.alpha432.oyvey.util.TextUtil;
-import net.minecraft.text.MutableText;
-import net.minecraft.text.Text;
-import net.minecraft.util.Formatting;
-
+import me.alpha432.oyvey.util.player.ChatUtil;
 
 public abstract class Command
         extends Feature {
@@ -26,22 +23,7 @@ public abstract class Command
     }
 
     public static void sendMessage(String message, Object... obj) {
-        sendMessage(TextUtil.text(message, obj));
-    }
-
-    public static void sendMessage(Text message) {
-        MutableText text = Text.empty();
-        text.append(OyVey.commandManager.getClientMessage() + " " + Formatting.GRAY);
-        text.append(message);
-        Command.sendSilentMessage(text);
-    }
-
-    public static void sendSilentMessage(Text message) {
-        if (Command.nullCheck()) {
-            return;
-        }
-        // TODO add silent support ig
-        mc.inGameHud.getChatHud().addMessage(message);
+        ChatUtil.sendMessage(TextUtil.text(message, obj));
     }
 
     public static String getCommandPrefix() {
