+++ src/main/java/me/alpha432/oyvey/features/modules/client/Notifications.java
@@ -1,0 +1,9 @@
+package me.alpha432.oyvey.features.modules.client;
+
+import me.alpha432.oyvey.features.modules.Module;
+
+public class Notifications extends Module {
+    public Notifications() {
+        super("Notifications", "Settings for notifications", Category.CLIENT);
+    }
+}
